QtRadio changes for binary header

+++ trunk/src/QtRadio/Spectrum.h
@@ -107,7 +107,7 @@ private:
     int moved;
 
     int sampleRate;
-    int meter;
+    short meter;
     int maxMeter;
     int meterCount;
     int subrx_meter;
